+++ src/Contract/SearchEngineInterface.php
@@ -4,17 +4,12 @@ declare(strict_types=1);
 
 namespace App\Contract;
 
+use App\DTO\SearchResult;
+
 interface SearchEngineInterface
 {
     /**
-     * @param array<string, mixed> $data
-     */
-    public function indexDocument(string $index, string $id, array $data): void;
-
-    /**
      * @param array<string, mixed> $query
-     *
-     * @return array<string, mixed>
      */
-    public function search(array $query): array;
+    public function search(array $query): SearchResult;
 }